+++ lib/ibm1097.h
@@ -93,7 +93,7 @@ static const unsigned short ibm1097_2uni[256] =
 
 
 static int
-ibm1097_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+ibm1097_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
    unsigned char c = *s;
    *pwc = (ucs4_t) ibm1097_2uni[c];
@@ -173,7 +173,7 @@ static const struct gap ibm1097_idx[] =
 
 
 static int
-ibm1097_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+ibm1097_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
    const struct gap* p2gap = ibm1097_idx;
    for (; p2gap->start < 0xffff ; p2gap++) {
