+++ lib/ibm1164.h
@@ -92,7 +92,7 @@ static const unsigned short ibm1164_2uni[256] =
 };
 
 static int
-ibm1164_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+ibm1164_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
    unsigned char c = *s;
    *pwc = (ucs4_t) ibm1164_2uni[c];
@@ -154,7 +154,7 @@ static const struct gap ibm1164_idx[] =
 };
 
 static int
-ibm1164_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+ibm1164_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
    const struct gap* p2gap = ibm1164_idx;
    for (; p2gap->start < 0xffff ; p2gap++) {
